+++ config/elasticsearch.php
@@ -46,6 +46,8 @@ return [
                     'scheme'     => env('ELASTICSEARCH_SCHEME', null),
                     'user'       => env('ELASTICSEARCH_USER', null),
                     'pass'       => env('ELASTICSEARCH_PASS', null),
+
+                    // If you are connecting to an Elasticsearch instance on AWS, you will need these values as well
                     'aws'        => env('AWS_ELASTICSEARCH_ENABLED', false),
                     'aws_region' => env('AWS_REGION', ''),
                     'aws_key'    => env('AWS_ACCESS_KEY_ID', ''),
@@ -172,19 +174,21 @@ return [
             /**
              * Endpoint
              *
-             * @see https://www.elastic.co/guide/en/elasticsearch/client/php-api/2.0/_configuration.html#_set_the_endpoint_closure
+             * @see https://www.elastic.co/guide/en/elasticsearch/client/php-api/6.0/_configuration.html#_set_the_endpoint_closure
              */
 
             'endpoint' => null,
 
 
             /**
-             * register additional namespaces
+             * Register additional namespaces
              *
-             * e.g.: for registerNamespace(XPack::Security()), 'namespaces' => XPack::Security()
-             *       or as an array: 'namespaces' => [XPack::Security(), XPack::Watcher()]
+             * An array of additional namespaces to register.
+             *
+             * @example 'namespaces' => [XPack::Security(), XPack::Watcher()]
+             * @see https://www.elastic.co/guide/en/elasticsearch/client/php-api/current/ElasticsearchPHP_Endpoints.html#Elasticsearch_ClientBuilderregisterNamespace_registerNamespace
              */
-            'namespaces' => null
+            'namespaces' => []
 
         ],